+++ app/src/main/java/com/devmobile/viajei/EntretenimentoActivity.java
@@ -42,7 +42,7 @@ public class EntretenimentoActivity extends AppCompatActivity {
 
         SharedPreferences sharedPreferences = PreferenceManager.getDefaultSharedPreferences(EntretenimentoActivity.this);
         String destino = sharedPreferences.getString("destino", "");
-        int idUsuario = sharedPreferences.getInt("idUsuario", -1);
+        long idUsuario = sharedPreferences.getLong("idUsuario", -1);
 
         TextView destinoTextView = findViewById(R.id.nome_destino);
         String textoDestino = getString(R.string.destino) + " " + destino;
